use zu in printf

+++ src/evenodd.h
@@ -182,8 +182,8 @@ struct depth_extend_sub_t {
                         number_lmc += (lmctype[t] > +LMC_EQUAL);
                 }
                 if (verbose) {
-                        myprintf ("lmc %ld/%ld\n", number_lmc, (long)lmctype.size ());
-                        myprintf ("valididx size %ld\n", (long)valididx.size ());
+                        myprintf ("lmc %ld/%zu\n", number_lmc, lmctype.size ());
+                        myprintf ("valididx size %zu\n", valididx.size ());
                 }
         }
 };
